address deletion handled in addresses context + references updated on deletion
